+++ pyrtl/rtllib/multipliers.py
@@ -292,7 +292,7 @@ def generalized_fma(
         for bit_loc, bit in enumerate(wire):
             bits[bit_loc].append(bit)
 
-    result_bitwidth = longest_wire_len + int(
-        math.ceil(math.log2(len(add_wires) + len(mult_pairs)))
+    result_bitwidth = longest_wire_len + math.ceil(
+        math.log2(len(add_wires) + len(mult_pairs))
     )
     return reducer(bits, result_bitwidth, adder_func)
